- create DependencyConfigFactory
